pyproject.toml
[tool.poetry]
name = "nonebot_plugin_farm"
version = "1.3.1"
description = "快乐的农场时光"
authors = ["Art_Sakura <1754798088@qq.com>"]
license = "GPL-3.0-or-later"
readme = "README.md"
repository = "https://github.com/Shu-Ying/nonebot_plugin_farm.git"

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = "^2.1.3"
aiosqlite = "^0.21.0"
httpx = ">=0.27.2,<1.0.0"
zhenxun-utils = "^0.2.1"
nonebot-plugin-alconna = ">=0.46.3,<1.0.0"
nonebot-plugin-uninfo = "^0.6.0"
nonebot-plugin-waiter = "^0.8.1"
nonebot_plugin_localstore = "^0.7.4"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
